+++ koala/cogs/text_filter/__init__.py
@@ -1,2 +1,2 @@
 from . import utils, db, models
-from .cog import TextFilter
+from .cog import TextFilter, setup