+++ koala/cogs/twitch_alert/__init__.py
@@ -1,2 +1,2 @@
 from . import utils, db, twitch_handler, log, models
-from .cog import TwitchAlert
+from .cog import TwitchAlert, setup