+++ koala/cogs/twitch_alert/cog.py
@@ -7,13 +7,11 @@ import re
 # Own modules
 import koalabot
 from koalabot import COMMAND_PREFIX as CP
-from koala.models import GuildExtensions
-from koala.db import session_manager, insert_extension
+from koala.db import insert_extension
 from koala.colours import KOALA_GREEN
 from koala.utils import error_embed, is_channel_in_guild
-from .core import create_team_alerts
+from . import core
 from .log import logger
-from .models import UserInTwitchAlert, TwitchAlerts, UserInTwitchTeam, TeamInTwitchAlert
 from .db import TwitchAlertDBManager
 from .utils import DEFAULT_MESSAGE, TWITCH_USERNAME_REGEX, \
     LOOP_CHECK_LIVE_DELAY, REFRESH_TEAMS_DELAY, TEAMS_LOOP_CHECK_LIVE_DELAY
@@ -22,7 +20,6 @@ from .env import TWITCH_KEY, TWITCH_SECRET
 # Libs
 import discord
 from discord.ext import commands, tasks
-from sqlalchemy import select, or_, delete, and_, update, null, func
 
 
 # Constants
@@ -186,7 +183,7 @@ class TwitchAlert(commands.Cog):
         channel_id = channel.id
         twitch_username = str.lower(twitch_username)
         if not re.search(TWITCH_USERNAME_REGEX, twitch_username):
-            raise discord.errors.InvalidArgument(
+            raise ValueError(
                 "The given twitch_username is not a valid username (please use lowercase)")
 
         # Check the channel specified is in this guild
@@ -278,7 +275,7 @@ class TwitchAlert(commands.Cog):
         team_name = str.lower(team_name)
 
         if not re.search(TWITCH_USERNAME_REGEX, team_name):
-            raise discord.errors.InvalidArgument(
+            raise ValueError(
                 "The given team_name is not a valid twitch team name (please use lowercase)")
 
         # Check the channel specified is in this guild
@@ -398,6 +395,7 @@ class TwitchAlert(commands.Cog):
         When the bot is started up, the loop begins
         :return:
         """
+        await self.ta_database_manager.setup_twitch_handler()
         if not self.running:
             self.start_loops()
 
@@ -420,108 +418,16 @@ class TwitchAlert(commands.Cog):
         sends alerts when online, removing them when offline
         :return:
         """
-        start = time.time()
-        # logger.info("TwitchAlert: User Loop Started")
-        sql_find_users = select(UserInTwitchAlert.twitch_username) \
-            .join(TwitchAlerts, UserInTwitchAlert.channel_id == TwitchAlerts.channel_id) \
-            .join(GuildExtensions, TwitchAlerts.guild_id == GuildExtensions.guild_id) \
-            .where(or_(GuildExtensions.extension_id == 'TwitchAlert', GuildExtensions.extension_id == 'All'))
-        # "SELECT twitch_username " \
-        #              "FROM UserInTwitchAlert " \
-        #              "JOIN TwitchAlerts TA on UserInTwitchAlert.channel_id = TA.channel_id " \
-        #              "JOIN (SELECT extension_id, guild_id FROM GuildExtensions " \
-        #              "WHERE extension_id = 'twitch_alert' OR extension_id = 'All') GE on TA.guild_id = GE.guild_id;"
-        with session_manager() as session:
-            users = session.execute(sql_find_users).all()
-
-        usernames = [str.lower(user.twitch_username) for user in users]
-
-        if not usernames:
-            return
-
-        user_streams = self.ta_database_manager.twitch_handler.get_streams_data(usernames)
-        if user_streams is None:
-            return
-
-        # Deals with online streams
-        for streams_details in user_streams:
-            try:
-                if streams_details.get('type') == "live":
-                    current_username = str.lower(streams_details.get("user_login"))
-                    old_len = len(usernames)
-                    usernames.remove(current_username)
-                    if len(usernames) == old_len:
-                        logger.error(f"TwitchAlert: {streams_details.get('user_login')} not found in the user list")
-
-                    sql_find_message_id = select(UserInTwitchAlert.channel_id,
-                                                 UserInTwitchAlert.message_id,
-                                                 UserInTwitchAlert.custom_message,
-                                                 TwitchAlerts.default_message) \
-                        .join(TwitchAlerts, UserInTwitchAlert.channel_id == TwitchAlerts.channel_id) \
-                        .join(GuildExtensions, TwitchAlerts.guild_id == GuildExtensions.guild_id) \
-                        .where(and_(and_(or_(GuildExtensions.extension_id == 'TwitchAlert',
-                                             GuildExtensions.extension_id == 'All'),
-                                         UserInTwitchAlert.twitch_username == current_username),
-                                    UserInTwitchAlert.message_id == null()))
-                    # "SELECT UserInTwitchAlert.channel_id, message_id, custom_message, default_message " \
-                    # "FROM UserInTwitchAlert " \
-                    # "JOIN TwitchAlerts TA on UserInTwitchAlert.channel_id = TA.channel_id " \
-                    # "JOIN (SELECT extension_id, guild_id FROM GuildExtensions " \
-                    # "WHERE extension_id = 'TwitchAlert' " \
-                    # "  OR extension_id = 'All') GE on TA.guild_id = GE.guild_id " \
-                    # "WHERE twitch_username = ?;"
-
-                    results = session.execute(sql_find_message_id).all()
-
-                    new_message_embed = None
-
-                    for result in results:
-                        channel_id = result.channel_id
-                        message_id = result.message_id
-                        custom_message = result.custom_message
-                        channel_default_message = result.default_message
-
-                        channel = self.bot.get_channel(id=channel_id)
-                        try:
-                            # If no Alert is posted
-                            if message_id is None:
-                                if new_message_embed is None:
-                                    if custom_message is not None:
-                                        message = custom_message
-                                    else:
-                                        message = channel_default_message
-
-                                    new_message_embed = await self.ta_database_manager.create_alert_embed(streams_details, message)
-
-                                if new_message_embed is not None and channel is not None:
-                                    new_message = await channel.send(embed=new_message_embed)
-                                    sql_update_message_id = update(UserInTwitchAlert).where(and_(
-                                        UserInTwitchAlert.channel_id == channel_id,
-                                        UserInTwitchAlert.twitch_username == current_username)) \
-                                        .values(message_id=new_message.id)
-                                    session.execute(sql_update_message_id)
-                                    session.commit()
-                        except discord.errors.Forbidden as err:
-                            logger.warning(f"TwitchAlert: {err}  Name: {channel} ID: {channel.id}")
-                            sql_remove_invalid_channel = delete(TwitchAlerts).where(
-                                TwitchAlerts.channel_id == channel.id)
-                            session.execute(sql_remove_invalid_channel)
-                            session.commit()
-
-            except Exception as err:
-                logger.error(f"TwitchAlert: User Loop error {err}")
-
-        # Deals with remaining offline streams
-        await self.ta_database_manager.delete_all_offline_streams(usernames)
-        time_diff = time.time() - start
-        if time_diff > 5:
-            logger.warning(f"TwitchAlert: User Loop Finished in > 5s | {time_diff}s")
+        try:
+            await core.create_user_alerts(self.bot, self.ta_database_manager)
+        except Exception as err:
+            logger.error("Twitch user live loop error: ", exc_info=err)
 
     @tasks.loop(minutes=REFRESH_TEAMS_DELAY)
     async def loop_update_teams(self):
         start = time.time()
         # logger.info("TwitchAlert: Started Update Teams")
-        self.ta_database_manager.update_all_teams_members()
+        await self.ta_database_manager.update_all_teams_members()
         time_diff = time.time() - start
         if time_diff > 5:
             logger.warning(f"TwitchAlert: Teams updated in > 5s | {time_diff}s")
@@ -534,11 +440,12 @@ class TwitchAlert(commands.Cog):
         :return:
         """
         try:
-            await create_team_alerts(self.bot, self.ta_database_manager)
+            await core.create_team_alerts(self.bot, self.ta_database_manager)
         except Exception as err:
             logger.error("Twitch team live loop error: ", exc_info=err)
 
-def setup(bot: koalabot) -> None:
+
+async def setup(bot: koalabot) -> None:
     """
     Load this cog to the KoalaBot.
     :param bot: the bot client for KoalaBot
@@ -547,5 +454,5 @@ def setup(bot: koalabot) -> None:
         logger.error("TwitchAlert not started. API keys not found in environment.")
         insert_extension("TwitchAlert", 0, False, False)
     else:
-        bot.add_cog(TwitchAlert(bot))
+        await bot.add_cog(TwitchAlert(bot))
         logger.info("TwitchAlert is ready.")