+++ tests/cogs/base/test_cog.py
@@ -14,13 +14,14 @@ import discord
 import discord.ext.test as dpytest
 import mock
 import pytest
+import pytest_asyncio
 from discord.ext import commands
 from sqlalchemy import delete
 
 # Own modules
 import koalabot
 from koala.cogs import BaseCog
-from koala.cogs.base.cog import setup as setup_cog
+from koala.cogs.base.cog import setup as setup_cog, BaseCog
 from koala.db import session_manager
 from koala.colours import KOALA_GREEN
 from koala.models import KoalaExtensions, GuildExtensions
@@ -38,27 +39,21 @@ def setup_is_dpytest():
     koalabot.is_dpytest = False
 
 
-# Test TwitchAlert
-@pytest.fixture(scope='function', autouse=True)
+@pytest_asyncio.fixture(scope='function', autouse=True)
 async def base_cog(bot: commands.Bot):
     """ setup any state specific to the execution of the given module."""
-    with session_manager() as session:
-        session.execute(delete(KoalaExtensions))
-        session.execute(delete(GuildExtensions))
-        session.commit()
-
-    base_cog = BaseCog(bot)
-    bot.add_cog(base_cog)
+    cog = BaseCog(bot)
+    await bot.add_cog(cog)
     await dpytest.empty_queue()
     dpytest.configure(bot)
-    return base_cog
+    return cog
 
 
-@mock.patch("koalabot.COGS_DIR", "tests/tests_utils/fake_load_all_cogs")
-@mock.patch("koalabot.ENABLED_COGS", [])
+@mock.patch("koalabot.COGS_PACKAGE", "tests.tests_utils.fake_load_all_cogs")
+@mock.patch("koalabot.ENABLED_COGS", ["greetings_cog"])
 @pytest.mark.asyncio
-async def test_list_koala_ext_disabled(base_cog):
-    koalabot.load_all_cogs()
+async def test_list_koala_ext_disabled(bot, base_cog):
+    await koalabot.load_all_cogs(bot)
     await dpytest.message(koalabot.COMMAND_PREFIX + "listExt")
     expected_embed = discord.Embed()
     expected_embed.title = "Enabled extensions"
@@ -68,11 +63,11 @@ async def test_list_koala_ext_disabled(base_cog):
     assert dpytest.verify().message().embed(embed=expected_embed)
 
 
-@mock.patch("koalabot.COGS_DIR", "tests/tests_utils/fake_load_all_cogs")
-@mock.patch("koalabot.ENABLED_COGS", [])
+@mock.patch("koalabot.COGS_PACKAGE", "tests.tests_utils.fake_load_all_cogs")
+@mock.patch("koalabot.ENABLED_COGS", ['greetings_cog'])
 @pytest.mark.asyncio
-async def test_enable_koala_ext(base_cog):
-    koalabot.load_all_cogs()
+async def test_enable_koala_ext(bot, base_cog):
+    await koalabot.load_all_cogs(bot)
     await dpytest.message(koalabot.COMMAND_PREFIX + "enableExt Greetings")
     expected_embed = discord.Embed()
     expected_embed.title = "Greetings enabled"
@@ -82,11 +77,11 @@ async def test_enable_koala_ext(base_cog):
     assert dpytest.verify().message().embed(embed=expected_embed)
 
 
-@mock.patch("koalabot.COGS_DIR", "tests/tests_utils/fake_load_all_cogs")
-@mock.patch("koalabot.ENABLED_COGS", [])
+@mock.patch("koalabot.COGS_PACKAGE", "tests.tests_utils.fake_load_all_cogs")
+@mock.patch("koalabot.ENABLED_COGS", ['greetings_cog'])
 @pytest.mark.asyncio
-async def test_disable_koala_ext(base_cog):
-    await test_enable_koala_ext(base_cog)
+async def test_disable_koala_ext(bot, base_cog):
+    await test_enable_koala_ext(bot, base_cog)
     await dpytest.message(koalabot.COMMAND_PREFIX + "disableExt Greetings")
     expected_embed = discord.Embed()
     expected_embed.title = "Greetings disabled"
@@ -191,8 +186,8 @@ async def test_invalid_clear(base_cog: BaseCog):
 @pytest.mark.asyncio
 async def test_load_cog(base_cog: BaseCog):
     with mock.patch.object(discord.ext.commands.bot.Bot, 'load_extension') as mock1:
-        await dpytest.message(koalabot.COMMAND_PREFIX + "load_cog BaseCog")
-    mock1.assert_called_with('koala.cogs.BaseCog')
+        await dpytest.message(koalabot.COMMAND_PREFIX + "load_cog base")
+    mock1.assert_called_with(".base", package="koala.cogs")
 
 
 @pytest.mark.asyncio
@@ -204,29 +199,28 @@ async def test_invalid_load_cog(base_cog: BaseCog):
 
 @pytest.mark.asyncio
 async def test_unload_base_cog(base_cog: BaseCog):
-    with mock.patch.object(discord.ext.commands.Context, 'send') as mock1:
+    with pytest.raises(discord.ext.commands.CommandInvokeError, match="Sorry, you can't unload the base cog"):
         await dpytest.message(koalabot.COMMAND_PREFIX + "unload_cog BaseCog")
-    mock1.assert_called_with("Sorry, you can't unload the base cog")
 
 
+@mock.patch("koalabot.COGS_PACKAGE", "tests.tests_utils.fake_load_all_cogs")
 @pytest.mark.asyncio
 async def test_load_valid_cog(base_cog: BaseCog):
-    base_cog.COGS_DIR = "tests/tests_utils/fake_load_all_cogs"
     with mock.patch.object(discord.ext.commands.bot.Bot, 'load_extension') as mock1:
         await dpytest.message(koalabot.COMMAND_PREFIX + "load_cog Greetings")
-    mock1.assert_called_with("tests.tests_utils.fake_load_all_cogs.Greetings")
+    mock1.assert_called_with(".Greetings", package="tests.tests_utils.fake_load_all_cogs")
 
 
+@mock.patch("koalabot.COGS_PACKAGE", "tests.tests_utils.fake_load_all_cogs")
 @pytest.mark.asyncio
 async def test_load_and_unload_valid_cog(base_cog: BaseCog):
-    base_cog.COGS_DIR = "tests/tests_utils/fake_load_all_cogs"
     with mock.patch.object(discord.ext.commands.bot.Bot, 'load_extension') as mock1:
         await dpytest.message(koalabot.COMMAND_PREFIX + "load_cog Greetings")
-    mock1.assert_called_with("tests.tests_utils.fake_load_all_cogs.Greetings")
+    mock1.assert_called_with(".Greetings", package="tests.tests_utils.fake_load_all_cogs")
 
     with mock.patch.object(discord.ext.commands.bot.Bot, 'unload_extension') as mock1:
         await dpytest.message(koalabot.COMMAND_PREFIX + "unload_cog Greetings")
-    mock1.assert_called_with('tests.tests_utils.fake_load_all_cogs.Greetings')
+    mock1.assert_called_with(".Greetings", package="tests.tests_utils.fake_load_all_cogs")
 
 
 @pytest.mark.asyncio
@@ -244,7 +238,7 @@ async def test_version(base_cog: BaseCog):
 
 
 @pytest.mark.asyncio
-async def test_setup():
+async def test_setup(bot):
     with mock.patch.object(discord.ext.commands.bot.Bot, 'add_cog') as mock1:
-        setup_cog(koalabot.bot)
+        await setup_cog(bot)
     mock1.assert_called()