+++ tests/test_utils.py
@@ -15,6 +15,7 @@ import discord
 import discord.ext.test as dpytest
 import mock
 import pytest
+import pytest_asyncio
 from discord.ext import commands
 
 # Own modules
@@ -55,7 +56,7 @@ def test_format_db_path_windows():
 
 @pytest.mark.parametrize("msg_content", [" ", "something"])
 @pytest.mark.asyncio
-async def test_wait_for_message_not_none(msg_content, utils_cog):
+async def test_wait_for_message_not_none(msg_content, utils_cog: LastCtxCog):
     await dpytest.message(koalabot.COMMAND_PREFIX + "store_ctx")
     ctx = utils_cog.get_last_ctx()
     config: dpytest.RunnerConfig = dpytest.get_config()
@@ -69,7 +70,7 @@ async def test_wait_for_message_not_none(msg_content, utils_cog):
 
 
 @pytest.mark.asyncio
-async def test_wait_for_message_none(utils_cog):
+async def test_wait_for_message_none(utils_cog: LastCtxCog):
     await dpytest.message(koalabot.COMMAND_PREFIX + "store_ctx")
     ctx: commands.Context = utils_cog.get_last_ctx()
     config: dpytest.RunnerConfig = dpytest.get_config()
@@ -78,10 +79,11 @@ async def test_wait_for_message_none(utils_cog):
     assert not msg
     assert channel == ctx.channel
 
-@pytest.fixture(autouse=True)
-def utils_cog(bot):
+
+@pytest_asyncio.fixture(autouse=True)
+async def utils_cog(bot: commands.Bot):
     utils_cog = LastCtxCog(bot)
-    bot.add_cog(utils_cog)
+    await bot.add_cog(utils_cog)
     dpytest.configure(bot)
     logger.info("Tests starting")
     return utils_cog