+++ CHANGELOG.md
@@ -9,6 +9,21 @@ administrators
 - made email verification case insensitive
 - Added optimisations for verified roles
 
+## [0.6.0] - 01-01-2023
+- Upgraded to discord.py 2.1.0
+
+## [0.5.9] - 13-07-2022
+### Verify
+- Fix an issue where reVerify would fail if run multiple times
+
+## [0.5.8] - 28-04-2022
+### TwitchAlert
+- Fix user sql error
+
+## [0.5.7] - 28-04-2022
+### TwitchAlert
+- Fix issue where users alerts would sometimes delete themselves 
+
 ## [0.5.6] - 19-04-2022
 ### TwitchAlert
 - Fix issue where duplicate team alerts cannot be removed
@@ -284,6 +299,10 @@ manually if a timer is deleted.
 - `removeProtectedRoleColour <role_str>` Removes a role, via ID, mention or name, from the list of protected roles.
 
 [Unreleased]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.1.0...HEAD
+[0.6.0]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.9...v0.6.0
+[0.5.9]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.8...v0.5.9
+[0.5.8]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.7...v0.5.8
+[0.5.7]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.6...v0.5.7
 [0.5.6]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.5...v0.5.6
 [0.5.5]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.4...v0.5.5
 [0.5.4]: https://github.com/KoalaBotUK/KoalaBot/compare/v0.5.3...v0.5.4